redefined models and added certificates and resume

+++ src/main/java/com/workify/auth/models/Education.java
@@ -1,5 +1,6 @@
 package com.workify.auth.models;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -16,7 +17,7 @@ public class Education {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
 
     private String institution;
     private String degree;
